fix: do not execute deck deletion for Heist sheet on non GM

+++ module/actor/documents/heist-actor.mjs
@@ -260,6 +260,10 @@ export class HeistActor extends BaseActor {
   }
 
   async #deleteDecks() {
+    if (!game.users.activeGM) {
+      return;
+    }
+
     await this.jackDeck?.delete();
     await this.jackPile?.delete();
     await this.jackTestHand?.delete();
